requirements.txt
asgiref==3.7.2
black==23.7.0
bleach==6.0.0
click==8.1.6
crispy-bootstrap5==0.6
diff-match-patch==20230430
Django==4.2
django-crispy-forms==1.14.0
django-import-export==4.1.0
django-jazzmin==2.5.0
django-summernote==0.8.20.0
et-xmlfile==1.1.0
flake8==6.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
openpyxl==3.1.5
packaging==23.1
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
tablib==3.5.0
tomli==2.0.1
typing-extensions==4.7.1
webencodings==0.5.1
